move useTrackMutedIndicator hook
