+++ packages/react/src/components/participant/TrackMutedIndicator.tsx
@@ -1,37 +1,14 @@
 import * as React from 'react';
 import { mergeProps } from '../../utils';
-import { setupTrackMutedIndicator } from '@livekit/components-core';
 import type { Participant, Track } from 'livekit-client';
-import { useEnsureParticipant } from '../../context';
 import { getSourceIcon } from '../../assets/icons/util';
-import { useObservableState } from '../../hooks/internal/useObservableState';
+import { useTrackMutedIndicator } from '../../hooks';
 
 /** @public */
 export interface TrackMutedIndicatorProps extends React.HTMLAttributes<HTMLDivElement> {
   source: Track.Source;
   participant?: Participant;
   show?: 'always' | 'muted' | 'unmuted';
-}
-
-/** @public */
-export interface UseTrackMutedIndicatorOptions {
-  participant?: Participant;
-}
-
-/** @public */
-export function useTrackMutedIndicator(
-  source: Track.Source,
-  options: UseTrackMutedIndicatorOptions = {},
-) {
-  const p = useEnsureParticipant(options.participant);
-  const { className, mediaMutedObserver } = React.useMemo(
-    () => setupTrackMutedIndicator(p, source),
-    [p, source],
-  );
-
-  const isMuted = useObservableState(mediaMutedObserver, !!p.getTrack(source)?.isMuted);
-
-  return { isMuted, className };
 }
 
 /**
